// Frontend/src/app/interfaces/task.ts
export interface Task{
    id: string;
    name: string;
    description: string;
    startDate: string;
    endDate: string;
    startTime: string;
    endTime: string;
    projectId: string;
    collaboratorId: string;
    collaboratorName: string;
    createdAt: string;
    updatedAt: string;
    deletedAt: string;
}